+++ dart/test/protocol/sentry_gpu_test.dart
@@ -6,7 +6,7 @@ void main() {
   final sentryGpu = SentryGpu(
       name: 'fixture-name',
       id: 1,
-      vendorId: 2,
+      vendorId: '2',
       vendorName: 'fixture-vendorName',
       memorySize: 3,
       apiType: 'fixture-apiType',
@@ -17,7 +17,7 @@ void main() {
   final sentryGpuJson = <String, dynamic>{
     'name': 'fixture-name',
     'id': 1,
-    'vendor_id': 2,
+    'vendor_id': '2',
     'vendor_name': 'fixture-vendorName',
     'memory_size': 3,
     'api_type': 'fixture-apiType',
@@ -63,7 +63,7 @@ void main() {
       final copy = data.copyWith(
         name: 'name1',
         id: 11,
-        vendorId: 22,
+        vendorId: '22',
         vendorName: 'vendorName1',
         memorySize: 33,
         apiType: 'apiType1',
@@ -74,7 +74,7 @@ void main() {
 
       expect('name1', copy.name);
       expect(11, copy.id);
-      expect(22, copy.vendorId);
+      expect('22', copy.vendorId);
       expect('vendorName1', copy.vendorName);
       expect(33, copy.memorySize);
       expect('apiType1', copy.apiType);